Refactored Core
Split Core from 400+ line class into size smaller files
