+++ Assets/Scripts/Core/Core.DataBaseInterface.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: bbccef1651089ef4ca0384c4bc0f44ae
+timeCreated: 1446958681
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
