+++ Assets/Scripts/Core/Core.GameHelperFunctions.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: cf214a383cd066f48861a066c458bdff
+timeCreated: 1446958681
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
