+++ Assets/Scripts/Core/Core.GuiInterface.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c9e832e16c87f9645995c108fa6104ec
+timeCreated: 1446958681
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
