+++ Assets/Scripts/Core/Core.UnityInterface.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: b280175a9509d7649821af3bd445fc7a
+timeCreated: 1446958681
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
